setting up heading queries
